refactor: remove circular dependencies in compiler
